game over updates to score history

+++ Server/server.py
@@ -104,7 +104,6 @@ def server_hints():
     is_new = request.args.get("new")
     fail = {"result": False, "data": "Database Connection lost"}
     if not (user or country):
-        print(user, country, "fail")
         fail["data"] = "invalid input"
         return fail
     try:
@@ -126,7 +125,6 @@ def server_hints():
     # query, keyword = get_from_option(option, country)
     hints_list = get_hints(country, amount, cursor=cursor)
     hints = []
-    print("get_hints", get_hints)
     born_count, died_count, rests_count = len(hints_list["born"]), len(hints_list["died"]), \
                                             len(hints_list["rests"])
     if born_count > 0:
@@ -153,7 +151,6 @@ def server_hints():
                 break
     if len(hints) == 0:
         return json.dumps(["No Available Hint"])
-    print(hints)
     return json.dumps(hints)
 
 
@@ -326,13 +323,25 @@ def delete_game(game_id=None):
     except Exception as e:
         print(e)
         return DatabaseError(e)
+    count = count_records("games", cursor=cursor, where=f"id={game_id}")
+    if count < 1:
+        return -1
     rows = letters_query = f"DELETE FROM game_letter WHERE gid={game_id}"
     delete_query(query=letters_query, cursor=cursor, to_commit=False)
     locations_query = f"DELETE FROM game_locations WHERE gid={game_id}"
     rows = delete_query(query=locations_query, cursor=cursor)
+    score_query = f"SELECT uid, current_score FROM games WHERE id={game_id}"
+    score = select_query(query=score_query, cursor=cursor, is_many=False)
+    if score:
+        now = datetime.datetime.now()
+        dt_string = now.strftime("%Y-%m-%d %H:%M:%S")
+        score_query = f"INSERT INTO score_history (uid, score, datetime) VALUES " \
+                      f"({score[0]}, {score[1]}, '{dt_string}'); "
+        rows = insert_query(query=score_query, cursor=cursor)
     game_query = f"DELETE FROM games WHERE id={game_id}"
     rows = delete_query(query=game_query, cursor=cursor)
     return str(rows)
+
 
 
 @app.route('/user_country')
@@ -391,7 +400,6 @@ def new_game():
     if gid:
         delete_game(gid[0])
     country = get_random_country(uid)
-    print(country)
     return country
 
 
@@ -543,7 +551,6 @@ def add_person():
     if name == '':
         return "-1"
     does_exist = select_query(query=f"SELECT id FROM people_info WHERE name='{name}'", is_many=False, cursor=cursor)
-    print(does_exist)
     if not does_exist:
         bornin = arg['bornin'] if 'bornin' in keys else None
         diedin = arg['diedin'] if 'diedin' in keys else None
